refactor: move ancestor loading from useEffect hook to navigation handler
Moved the responsibility of loading missing ancestors from a useEffect hook
with costly item dependencies to the navigation handler. This is more efficient
as it only runs when navigation occurs and ancestors are actually missing,
rather than re-running every time any item in the cache changes.

- Removed useLoadAncestors hook that had expensive dependencies
- Added ancestor loading logic to navigation handler
- Ancestors are now loaded automatically when navigating to a sub-tile
- More efficient and architecturally sound approach
