+++ src/app/map/Cache/Handlers/navigation-handler.ts
@@ -3,11 +3,14 @@ import type { CacheAction, CacheState } from "../State/types";
 import { cacheActions } from "../State/actions";
 import type { DataOperations, NavigationOperations } from "./types";
 import { CoordSystem } from "~/lib/domains/mapping/utils/hex-coordinates";
+import type { TileData } from "../../types/tile-data";
+import type { ServerService } from "../Services/types";
 
 export interface NavigationHandlerConfig {
   dispatch: React.Dispatch<CacheAction>;
   getState: () => CacheState;
   dataHandler: DataOperations;
+  serverService?: ServerService;
   // For testing, we can inject these dependencies
   router?: {
     push: (url: string) => void;
@@ -151,6 +154,82 @@ export function createNavigationHandler(config: NavigationHandlerConfig) {
         dataHandler.prefetchRegion(itemCoordId).catch(error => {
           console.error('[NAV] Background region load failed:', error);
         });
+      }
+      
+      // 6. Check if ancestors need to be loaded
+      const centerItem = getState().itemsById[itemCoordId];
+      if (centerItem && centerItem.metadata.coordinates.path.length > 0) {
+        // Check if we have all ancestors by walking up the tree
+        let currentCoordId = itemCoordId;
+        let hasAllAncestors = true;
+        const missingAncestorLevels: string[] = [];
+        
+        while (true) {
+          const parentCoordId = CoordSystem.getParentCoordFromId(currentCoordId);
+          if (!parentCoordId) break; // Reached root
+          
+          if (!getState().itemsById[parentCoordId]) {
+            hasAllAncestors = false;
+            missingAncestorLevels.push(parentCoordId);
+          }
+          currentCoordId = parentCoordId;
+        }
+        
+        // Load ancestors if missing
+        if (!hasAllAncestors && centerItem.metadata.dbId) {
+          const loadAncestors = async () => {
+            try {
+              const centerDbId = parseInt(centerItem.metadata.dbId);
+              if (!isNaN(centerDbId) && config.serverService) {
+                const ancestors = await config.serverService.getAncestors(centerDbId);
+                
+                if (ancestors.length > 0) {
+                  // Convert to TileData format
+                  const ancestorItems: Record<string, TileData> = {};
+                  
+                  ancestors.forEach(ancestor => {
+                    const coordId = ancestor.coordinates;
+                    const coords = CoordSystem.parseId(coordId);
+                    
+                    ancestorItems[coordId] = {
+                      data: {
+                        name: ancestor.name,
+                        description: ancestor.descr,
+                        url: ancestor.url,
+                        color: '#000000', // Default color
+                      },
+                      metadata: {
+                        coordId,
+                        dbId: ancestor.id,
+                        depth: coords.path.length,
+                        parentId: ancestor.parentId ? ancestor.parentId.toString() : undefined,
+                        coordinates: coords,
+                        ownerId: ancestor.ownerId,
+                      },
+                      state: {
+                        isDragged: false,
+                        isHovered: false,
+                        isSelected: false,
+                        isExpanded: false,
+                        isDragOver: false,
+                        isHovering: false,
+                      },
+                    };
+                  });
+                  
+                  // Dispatch ancestors to cache
+                  dispatch(cacheActions.updateItems(ancestorItems));
+                  console.log('[NAV] Loaded', ancestors.length, 'ancestors for', itemCoordId);
+                }
+              }
+            } catch (error) {
+              console.error('[NAV] Failed to load ancestors:', error);
+            }
+          };
+          
+          // Load ancestors in background
+          void loadAncestors();
+        }
       }
 
       return {
@@ -321,6 +400,7 @@ export function useNavigationHandler(
   dispatch: React.Dispatch<CacheAction>,
   getState: () => CacheState,
   dataHandler: DataOperations,
+  serverService?: ServerService,
 ) {
   // Always call hooks unconditionally
   const router = useRouter();
@@ -331,6 +411,7 @@ export function useNavigationHandler(
     dispatch,
     getState,
     dataHandler,
+    serverService,
     router,
     searchParams,
     pathname,
